config/requirements.txt
django==4.2.7
djangorestframework==3.14.0
google-api-python-client==2.108.0
uwsgi==2.0.23
celery==5.3.5
redis==5.0.1